worker/requirements.txt
flask==3.0.0
ydb==3.23.0
boto3==1.34.34
yandex-cloud-ml-sdk==0.2.0
requests>=2.32.3
reportlab==4.0.9
ffmpeg-python==0.2.0
waitress==2.1.2
